+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     * Chain Lightning now respects the CL 20 cap for its damage dice
     * Destructive Dispel now calculates the DC based on the effective CL of the dispel and the highest mental stat to better support edge cases. Formula is 10 + 1/2 CL + Highest Mental Stat.
     * Expanded Arsenal can no longer be used to stack spell focus feats to increase DC
+    * Geniekind no longer causes too many damage instances
     * Magus spell combat/strike is now properly restricted to magus spells instead of the magus spell list
     * More of the unique metamagic rods will no longer default to active
     * Smite Evil/Smite Chaos/Mark of Justice attack bonus no longer stacks